compile typefix

+++ src/xBRZ/xbrz.cpp
@@ -217,9 +217,9 @@ double distYCbCrBuffered(uint32_t pix1, uint32_t pix2)
     const int g_diff = static_cast<int>(getGreen(pix1)) - getGreen(pix2);
     const int b_diff = static_cast<int>(getBlue (pix1)) - getBlue (pix2);
 
-    return diffToDist[(((r_diff + 0xFF) / 2) << 16) | //slightly reduce precision (division by 2) to squeeze value into single byte
-                                        (((g_diff + 0xFF) / 2) <<  8) |
-                                        (( b_diff + 0xFF) / 2)];
+    return diffToDist[size_t((((r_diff + 0xFF) / 2) << 16) | //slightly reduce precision (division by 2) to squeeze value into single byte
+                             (((g_diff + 0xFF) / 2) <<  8) |
+                             (((b_diff + 0xFF) / 2)      ))];
 #else //not noticeably faster:
     const int r_diff_tmp = ((pix1 & 0xFF0000) + 0xFF0000 - (pix2 & 0xFF0000)) / 2;
     const int g_diff_tmp = ((pix1 & 0x00FF00) + 0x00FF00 - (pix2 & 0x00FF00)) / 2; //slightly reduce precision (division by 2) to squeeze value into single byte
@@ -1157,8 +1157,8 @@ bool xbrz::equalColorTest(uint32_t col1, uint32_t col2, ColorFormat colFmt, doub
 void xbrz::bilinearScale(const uint32_t* src, int srcWidth, int srcHeight,
                          /**/  uint32_t* trg, int trgWidth, int trgHeight)
 {
-    bilinearScale(src, srcWidth, srcHeight, srcWidth * sizeof(uint32_t),
-                  trg, trgWidth, trgHeight, trgWidth * sizeof(uint32_t),
+    bilinearScale(src, srcWidth, srcHeight, srcWidth * (int)sizeof(uint32_t),
+                  trg, trgWidth, trgHeight, trgWidth * (int)sizeof(uint32_t),
     0, trgHeight, [](uint32_t pix) { return pix; });
 }
 
@@ -1166,8 +1166,8 @@ void xbrz::bilinearScale(const uint32_t* src, int srcWidth, int srcHeight,
 void xbrz::nearestNeighborScale(const uint32_t* src, int srcWidth, int srcHeight,
                                 /**/  uint32_t* trg, int trgWidth, int trgHeight)
 {
-    nearestNeighborScale(src, srcWidth, srcHeight, srcWidth * sizeof(uint32_t),
-                         trg, trgWidth, trgHeight, trgWidth * sizeof(uint32_t),
+    nearestNeighborScale(src, srcWidth, srcHeight, srcWidth * (int)sizeof(uint32_t),
+                         trg, trgWidth, trgHeight, trgWidth * (int)sizeof(uint32_t),
     0, trgHeight, [](uint32_t pix) { return pix; });
 }
